Add an option to disable the %n specifier.
The `%n` specifier writes the number of characters already written to
the address stored in the corresponding argument. This is rarely used
for legitimate purposes, and can be used as a write primitive if an
attacker gains control over the contents of a format string (see
https://www.owasp.org/index.php/Format_string_attack). Unless required
for backward compatibility with certain particular applications, `%n`
can generally be made into a no-op as a hardening measure.