Add RDRAND support to rumpkernel.
NetBSD's entropy system uses the RDRAND instruction on x86, but this
feature is #ifdef-ed out when NetBSD is built as a rumpkernel. Since
RDRAND is not a privileged instruction, it will in fact still work in
ring 3, so there is no reason not to use it on rumpkernels as well.