Use RDRAND to generate the stack protector value.
Previously, the value of the stack protector would be generated using
NetBSD's sysctl API, which does not become available until quite late in
the boot process, when several stack frames have already been
initialized with the previous canary value (usually 0). This commit
changes the stack protector init process so that it gets the value using
the RDRAND instruction, which is available on all modern x86 CPUs.
If RDRAND is not supported, the value be set to the terminator canary.

In addition to the stack canary, separate guard values are added for use
in page and malloc chunk headers, and these values are placed on their
own page and made read-only.