Add Xen callback function prototypes.
In Rumprun, the rumpkernel- and Xen-related sections of the codebase
cannot directly "see" each other, but rumpkernel needs to get a function
pointer to a Xen callback in order to implement `mprotect` and `mmap`.
This is the closest point in the code that I could find whose
definitions are visible to both relevant parts of the codebase.